create validate_utils file to fix circular imports
